# docs/openapi.yaml
openapi: 3.0.0
info:
  title: HotelHub API
  description: REST API for Hospitality Management System
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/service-requests:
    get:
      summary: Get all guest service requests
      responses:
        '200':
          description: List of service requests
    post:
      summary: Create a guest service request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [guestId, roomNumber, category]
              properties:
                guestId:
                  type: string
                roomNumber:
                  type: string
                category:
                  type: string
                  enum: [HOUSEKEEPING, ROOM_SERVICE, MAINTENANCE]
                specialInstructions:
                  type: string
      responses:
        '201':
          description: Service request created with estimated completion time and staff notification placeholder
        '400':
          description: Invalid input
        '422':
          description: Validation error
  /api/service-requests/{request_id}:
    get:
      summary: Get guest service request by ID
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service request found
        '404':
          description: Service request not found